1. 調整 Veggies 命名空間下的類別移動至 Models.Veggies 命名空間下，並調整相關參考的程式碼。 2. 格式化與清理 Veggie.cs 文件並增加 XML 註解。 3. 調整 Veggie.Name 屬性為 protected set。 4. 格式化與清理 Tomato.cs 文件並增加 XML 註解。 5. 格式化與清理 Onion.cs 文件並增加 XML 註解。
