+++ NineYi.BurgerShop/Burgers/Burger.cs
@@ -1,6 +1,6 @@
 ﻿using NineYi.BurgerShop.Models.Breads;
 using NineYi.BurgerShop.Meats;
-using NineYi.BurgerShop.Veggies;
+using NineYi.BurgerShop.Models.Veggies;
 using System;
 using System.Collections.Generic;
 using System.Linq;
