+++ NineYi.BurgerShop/Models/Veggies/Onion.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Veggies
+{
+    /// <summary>
+    /// 洋蔥。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Veggies.Veggie" />
+    public class Onion : Veggie
+    {
+        /// <summary>
+        /// 初始化 <see cref="Onion"/> 類別新的執行個體。
+        /// </summary>
+        public Onion()
+        {
+            this.Name = "Onion";
+        }
+    }
+}
