+++ NineYi.BurgerShop/Models/Veggies/Tomato.cs
@@ -1,0 +1,17 @@
+﻿namespace NineYi.BurgerShop.Models.Veggies
+{
+    /// <summary>
+    /// 番茄。
+    /// </summary>
+    /// <seealso cref="NineYi.BurgerShop.Models.Veggies.Veggie" />
+    public class Tomato : Veggie
+    {
+        /// <summary>
+        /// 初始化 <see cref="Tomato"/> 類別新的執行個體。
+        /// </summary>
+        public Tomato()
+        {
+            this.Name = "Tomato";
+        }
+    }
+}
