+++ NineYi.BurgerShop/NineYi.BurgerShop.csproj
@@ -58,9 +58,9 @@
     <Compile Include="Meats\Turkey.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Veggies\Onion.cs" />
-    <Compile Include="Veggies\Tomato.cs" />
-    <Compile Include="Veggies\Veggie.cs" />
+    <Compile Include="Models\Veggies\Onion.cs" />
+    <Compile Include="Models\Veggies\Tomato.cs" />
+    <Compile Include="Models\Veggies\Veggie.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
