--- NineYi.BurgerShop/Veggies/Onion.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Veggies
-{
-    /// <summary>
-    /// 洋蔥
-    /// </summary>
-    public class Onion : Veggie
-    {
-        public Onion()
-        {
-            this.Name = "Onion";
-        }
-    }
-}
