--- NineYi.BurgerShop/Veggies/Tomato.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NineYi.BurgerShop.Veggies
-{
-    /// <summary>
-    /// 番茄
-    /// </summary>
-    public class Tomato : Veggie
-    {
-        public Tomato()
-        {
-            this.Name = "Tomato";
-        }
-    }
-}
